--- a/include/GafferImage/ChannelDataProcessor.h
+++ b/include/GafferImage/ChannelDataProcessor.h
@@ -52,7 +52,7 @@ class GAFFERIMAGE_API ChannelDataProcessor : public ImageProcessor
 
 	public :
 
-		ChannelDataProcessor( const std::string &name=defaultName<ChannelDataProcessor>() );
+		ChannelDataProcessor( const std::string &name=defaultName<ChannelDataProcessor>(), bool premultiplyPlug = false );
 		~ChannelDataProcessor() override;
 
 		GAFFER_GRAPHCOMPONENT_DECLARE_TYPE( GafferImage::ChannelDataProcessor, ChannelDataProcessorTypeId, ImageProcessor );
@@ -61,6 +61,9 @@ class GAFFERIMAGE_API ChannelDataProcessor : public ImageProcessor
 
 		Gaffer::StringPlug *channelsPlug();
 		const Gaffer::StringPlug *channelsPlug() const;
+
+		Gaffer::BoolPlug *processUnpremultipliedPlug();
+		const Gaffer::BoolPlug *processUnpremultipliedPlug() const;
 
 	protected :
 
@@ -79,7 +82,10 @@ class GAFFERIMAGE_API ChannelDataProcessor : public ImageProcessor
 		/// @param outData The tile where the result of the operation should be written. It is initialized with the coresponding tile data from inPlug() which should be used as the input data.
 		virtual void processChannelData( const Gaffer::Context *context, const ImagePlug *parent, const std::string &channel, IECore::FloatVectorDataPtr outData ) const = 0;
 
+		void hashChannelData( const GafferImage::ImagePlug *output, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
+
 	private :
+		bool m_hasUnpremultPlug;
 
 		static size_t g_firstPlugIndex;