--- a/include/GafferImage/ColorProcessor.h
+++ b/include/GafferImage/ColorProcessor.h
@@ -56,6 +56,9 @@ class GAFFERIMAGE_API ColorProcessor : public ImageProcessor
 
 		GAFFER_GRAPHCOMPONENT_DECLARE_TYPE( GafferImage::ColorProcessor, ColorProcessorTypeId, ImageProcessor );
 
+		Gaffer::BoolPlug *processUnpremultipliedPlug();
+		const Gaffer::BoolPlug *processUnpremultipliedPlug() const;
+
 		Gaffer::StringPlug *channelsPlug();
 		const Gaffer::StringPlug *channelsPlug() const;