Once upon a time, long, long, ago, @HughMacdonald put a lot of work into adding deep support into Gaffer.  Unfortunately, the specific way that we were using an OIIO cache for input images didn't support deep ( in addition to conflicting with Gaffer's built in cache anyway ).  By the time we fixed this, everyone involved was focused on other priorities, and this PR kind of fell by the wayside.

I've finally gotten a chance to get this working with modern Gaffer, and add support for reading and writing deeps, in addition to going through all this code and trying to make sure it's a decent foundation for building some deep support that will hopefully perform reasonably well ( there are currently some performance issues with how it interacts with Gaffer's caching architecture, but I won't dive into that quite yet ).

This PR is not ready to merge quite yet, there's still some sawdust, but everything up until around DeepState should be ready for review - though note that github shows the commits all out of order.  I've taken a commenting pass on DeepState, which is really the important one to explain because that's where all the complexity is, but DeepMerge could probably use a few more comments.

Known outstanding issues:
* Reader and Writer code could probably use a cleanup / commenting pass
* Various TODOs still in the code
* Need to correctly handle OpenEXR DeepImageState header metadata
* Should create a DeepTidy python node to simplify the UI of DeepState when used for tidying.  Possibly hide DeepState from menu, and just show DeepTidy and Flatten.
* DeepRecolor is the last node needed to add the functionality needed to do a basic slapcomp using deep holdouts ( DeepHoldout is pretty key as well, but isn't strictly required ).