@danieldresser-ie 2019-10-17
Couple of cleanups to DeepMerge and ImageWriter.  As requested, I won't squash them in until you've finished your review.

Didn't quite finalize the fixes and tests for CollectImages, that will be next, and should get the tests all passing.