@danieldresser-ie 2019-10-18
Cool, that all looks pretty straightforward so far.

> The only real topic of conversation is those few FlatImageProcessors that seem to be trying to specialise hashDeepState(), and I'm inclined to say we should just standardise on "FlatImageProcessors always throw".

Sounds good to me, I just hadn't touched this stuff, but it does seem like unnecessary complexity.


One question though:  I think I do feel moderately strongly about keeping allPixelData() alongside image() and imageHash().  If you want to move image() and imageHash() to ImageAlgo anyway, I might as well take a swing at doing that now - I started playing with it, and it looks pretty straightforward.  Only question - what should these methods be named?  Are you happy with ImageAlgo::image and ImageAlgo::imageHash, or is that too non-descript?