@danieldresser-ie 2019-10-19
OK, I've now hit the majority of your feedback, plus some of my own tidying, plus the CollectImages changes to allow sample-matched deep, and squashed it together.  With the CollectImages changes, we should now be seeing the tests pass, which is nice.

3 bits of feedback I haven't hit yet:
* Removing Python wrapper for a FlatImageProcessor : This gave me a link error when importing _GafferImage.so : `RuntimeError: extension class wrapper for base class GafferImage::FlatImageProcessor has not been created yet`.  I'll think about whether this makes sense on Monday, but wanted to get the rest up for the moment.
* OSLImage support deep properly: you're probably right that this should be fairly straightforward.  Definitely seems worth doing.  Next on my list
* Mix : This is currently doing weird Deep passthrough hackery - if it continues inheriting from FlatImageProcessor I will just rip this out.  But thinking about how you suggest I handle CollectImages: "support Deep but only if the sample counts match exactly" - it would actually be quite useful if Mix did this as well.  Should be quite easy to write, and combined with OSLImage would allow you to do things like "apply the grade only within this sphere" which Lucien is interested in.