@danieldresser-ie 2019-10-21
Forgot to mention, two of your pieces of feedback which I addressed, but could still be worth commenting on:
* discardZeroAlpha : This was just moved to DeepState ( the part you really want on DeepMerge is discarding occluded, since that's the problem the merge can create.  But discarding occluded requires sorting, and I didn't want to reimplement too much of DeepState. ) It was left undecided whether to put an internal DeepState inside the DeepMerge we expose to users.
* enabled() function on DeepMerge: I've updated this to match the behaviour of Merge - enabled() function is removed, logic to pass through hash when only one input is connected has not been relocated elsewhere, tests for hash passthrough are now expected failures.  We can clean this up at the same time if we do so for Merge.  To be honest, I'm not sure if failing to optimize the case where there is just one input bothers me - I can't think when this would happen unless a user explicitly builds a graph with a useless Merge.