@johnhaddon 2019-10-22
> Removing Python wrapper for a FlatImageProcessor : This gave me a link error when importing _GafferImage.so

Did you remove the binding entirely? We definitely need a binding still, but I don't _think_ we need to use ComputeNodeWrapper for it.

> Mix : This is currently doing weird Deep passthrough hackery - if it continues inheriting from FlatImageProcessor I will just rip this out. But thinking about how you suggest I handle CollectImages: "support Deep but only if the sample counts match exactly" - it would actually be quite useful if Mix did this as well. Should be quite easy to write, and combined with OSLImage would allow you to do things like "apply the grade only within this sphere" which Lucien is interested in.

Sounds good. Presumably the mask input could be flat even, as long as the two main inputs had the same sample counts? I'd like to focus on getting to a mergeable point quickly though, so if this distracts from that then I'm happy to leave it as followup work.