@johnhaddon 2019-10-31
Here's the diff needed to get this building with Clang (in this case, on MacOS) :

```
diff --git a/include/GafferImage/CollectImages.h b/include/GafferImage/CollectImages.h
index 028861dcf..07644dd95 100644
--- a/include/GafferImage/CollectImages.h
+++ b/include/GafferImage/CollectImages.h
@@ -73,10 +73,10 @@ class GAFFERIMAGE_API CollectImages : public ImageProcessor
                void hashDataWindow( const GafferImage::ImagePlug *output, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
                Imath::Box2i computeDataWindow( const Gaffer::Context *context, const ImagePlug *parent ) const override;
 
-               virtual void hashDeepState( const GafferImage::ImagePlug *parent, const Gaffer::Context *context, IECore::MurmurHash &h ) const;
+               void hashDeepState( const GafferImage::ImagePlug *parent, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
                int computeDeepState( const Gaffer::Context *context, const ImagePlug *parent ) const override;
 
-               virtual void hashSampleOffsets( const GafferImage::ImagePlug *parent, const Gaffer::Context *context, IECore::MurmurHash &h ) const;
+               void hashSampleOffsets( const GafferImage::ImagePlug *parent, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
                IECore::ConstIntVectorDataPtr computeSampleOffsets( const Imath::V2i &tileOrigin, const Gaffer::Context *context, const ImagePlug *parent ) const override;
 
                void hashChannelNames( const GafferImage::ImagePlug *parent, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
diff --git a/include/GafferImage/Mix.h b/include/GafferImage/Mix.h
index 9b1ab82c5..93eb091b8 100644
--- a/include/GafferImage/Mix.h
+++ b/include/GafferImage/Mix.h
@@ -77,9 +77,8 @@ class GAFFERIMAGE_API Mix : public FlatImageProcessor
                void hashDataWindow( const GafferImage::ImagePlug *output, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
                void hashChannelNames( const GafferImage::ImagePlug *output, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
                void hashChannelData( const GafferImage::ImagePlug *output, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
-               virtual void hashDeepState( const GafferImage::ImagePlug *parent, const Gaffer::Context *context, IECore::MurmurHash &h ) const;
-               virtual void hashSampleOffsets( const GafferImage::ImagePlug *parent, const Gaffer::Context *context, IECore::MurmurHash &h ) const;
-
+               void hashDeepState( const GafferImage::ImagePlug *parent, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
+               void hashSampleOffsets( const GafferImage::ImagePlug *parent, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
 
                /// Sets the data window to the union of all of the data windows.
                Imath::Box2i computeDataWindow( const Gaffer::Context *context, const ImagePlug *parent ) const override;

```