@danieldresser-ie 2019-12-03
OK, to the best of my knowledge, this is now ready to merge.

Did some quick 2D perf tests:

2D slapcomp, basic operations on 3K production images:
Before: 1.9s
After: 1.9s

Denoiser Test:
Before: 8.5s
After ( without parallel access pattern change ): 8.5s
After ( with parallel access pattern change ): 5.7s